LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/position-area-scrolling-007-expected.txt


PASS initial position
FAIL Scroll down to 0,10 assert_equals: expected 250 but got 0
FAIL Scroll right to 10,10 assert_equals: expected 250 but got 0
FAIL Scroll up to 10,0 assert_equals: expected 250 but got 0
FAIL Scroll down to 10,20 assert_equals: expected 10 but got 0
FAIL Scroll down to 101,20 assert_equals: expected 250 but got 0
FAIL Scroll down to 101,30 assert_equals: expected 250 but got 0
